Add a trivial web app for viewing the temperature data in a chart.
In the pico probe make an outgoing queue so that readings aren't lost if
the uploader activates less frequently than the temperature reader
(and if there is a connection problem).  Also don't try connecting
forever as that get us stuck. Also collect garbage after the upload to
lessen the chance of running out of memory.