Get the tests working for sensor_reading such as they are after a huge amount of upheaval.
Big changes include how the SampleMethod enum is created and used everywhere.
We try to use set[SampleMethod] everywhere rather than set[str] so that
we are dealing with something that has been validated

We can use an in memory database for testing so it's quicker and that lets us drop
the whole thing and rebuild it fairly quickly which is simpler than adding data then
trying to remove it later.