Some updates to the app to make the display slightly more reasonable. The main attempt is to adjust the start_timestamp to boundaries that "make sense" e.g. for the day view the start boundary should be the start of that day.
In the week view we should have ticks that are just the date - no need
for a time.

It's still very imperfect esp when changing views from zoomed in (e.g.
hour) to zoomed out (e.g. day). The logic for this allows a graph that
is not a sensible multiple.