Try to make the selection of start and end time more reasonable.
If startTime locks onto the nearest start time, then endTime must not
also do so otherwise the default view will not contain the current partial hour,
day, or week.

Add a script for deploying.

Remove a dud alembic revision that we're not using.
update the index to be non generic and the readme a bit too.